requirements.txt
pandas==2.2.2
numpy==2.0.1
pyarrow
paho-mqtt==2.1.0
python-dateutil

